Remove multiple fetchers
